apps/producer/requirements.txt
kafka-python==2.0.2
orjson==3.10.7
